Move task function to generator
